# vo_tct: further optimizations