Tested and benchmarked with every option, here are the results:

| vo-tct-algo | vo-tct-256 | frames dropped old | frames dropped new |
| -------------- | ------------- | --------------------- | ------------- |
| half-blocks | no | 1343 |  1000 | 
| half-blocks | yes | 1012 | 567 |
| plain | no | 832 | 108 |
| plain | yes | 360 | 20 |
-------
On more throughput intensive settings (half-blocks, truecolor), the biggest bottleneck is the terminal emulator itself.

Things change when we get to less throughput intensive ones. There we can see a true CPU bottleneck that has been resolved by not calling `fwrite()`/`printf()` for every pixel.