@Cloud11665 2021-08-29
Trying to eliminate the CPU bottleneck, i have replaced all of the `sprintf()`s with `snprintf()`s, and added `prefix_len` argument to the `sprint_seq{1,3}()` functions. Here are the results:

| vo-tct-algo | vo-tct-256 | frames dropped  |
| -------------- | ------------- | ------------------ |
| half-blocks | no | 901 | 
| half-blocks | yes | 430 |
| plain | no | 16 |
| plain | yes | 1 |
---------
Another optimization, could be not checking `if (term256)` for every pixel, but for every frame.
The only downside, would be moderate amounts of boilerplate code