@Cloud11665 2021-08-29
Since vt codes change the global state of the terminal, we are wasting throughput by printing out the same escape code.

Why print `\033[48;5;192;100;9m\033[38;5;10;40;16m▄\033[48;5;200;9;180m\033[38;5;10;40;16m▄`, when we could just print `\033[48;5;192;100;9m\033[38;5;10;40;16m▄\033[48;5;200;9;180m▄` (since the second fg code was the same as the first one)
saving us 22% of throughput.
Here are the benchmark results:

| vo-tct-algo | vo-tct-256 | old method  | fbuff | fbuff + sn | fbuff + sn + previous escape code  |
| ----------- | ---------- | ----------- | ----------- | ---------------------- | ----------------------------------- |
| half-blocks | no         | 1343        | 1000        | 901                    | 228                                 |
| half-blocks | yes        | 1012        | 567         | 430                    | 33                                  |
| plain       | no         | 832         | 108         | 16                     | 2                                   |
| plain       | yes        | 360         | 20          | 1                      | 0                                   |